reduce public interface of c func files
