+++ src/func.h
@@ -1,10 +1,7 @@
 #ifndef GIG_FUNC_H
 #define GIG_FUNC_H
-#include "func/gig_arg_map.h"
-#include "func/gig_argument.h"
-#include "func/gig_callback.h"
-#include "func/gig_data_type.h"
-#include "func/gig_function.h"
-#include "func/gig_function_private.h"
-#include "func/gig_util.h"
+#include "func/gig_arg_map_pub.h"
+#include "func/gig_callback_pub.h"
+#include "func/gig_function_pub.h"
+#include "func/gig_util_pub.h"
 #endif
