+++ src/func/gig_callback_pub.h
@@ -13,16 +13,10 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-#ifndef GIG_CALLBACK_H
-#define GIG_CALLBACK_H
+#ifndef GIG_CALLBACK_PUB_H
+#define GIG_CALLBACK_PUB_H
 
-#include <girepository.h>
-#include <libguile.h>
-#include "gig_callback_pub.h"
 #include "../core.h"
-
-SCM gig_callback_to_scm(const char *name, GICallbackInfo *info, void *proc);
-void *gig_callback_to_c(const char *name, GICallbackInfo *callback_info, SCM s_func);
 
 GIG_API void gig_init_callback(void);
